Only run the tests in GA on pushes to the master branch
Otherwise we get double runs when pushing to another branch with a PR
already open. I forgot about this scenario when setting it up before,
but I've seen it in a couple other projects.

This unfortunately means that simple pushes to a different branch won't
build, until you've opened a PR. But you can always open a draft PR to
get the build going there then, which also helps to have a place to
document and/or discuss stuff as you're working through it.