Revert GH links in Changelog to original Rails Autoscale repo
SHA commit references are fine since they carry over to the new repo,
however PRs and issues do not, and they reset. (start over from 1.) This
means we can't just update those, we need to either keep pointing them
to the original repo (as I'm doing here), or point them all to specific
SHAs so they work on this repo.

I'm going to revert this and point back to the original repo since
that's not going anywhere any time soon, and we can keep the references
and connection explicit.

The only one I'm keeping pointing to this repo is the "unreleased"
compare link.

[ci skip]