Apply StandardRB to the code
This uses the command line `standardrb --fix` to automatically fix most
of the offenders.

Most of the changes are related to single => double quotes, but there's
a few other alignment & syntax changes here and there.

There's still a few that will need to be fixed manually, I'll commit
those separately.