Run the StandardRB linter with the test build
This should help catch formatting issues during PRs & reviews.

I could not get the GH Action ones to work, trying to force a lint
failure didn't trigger a failure with them (which was supposed to fail
and annotate the problems in the PR). We can circle back on it in the
future, but for now this should be enough to help us catch issues
upfront.

For reference, these are the two I tried:
https://github.com/andrewmcodes/standardrb-action
https://github.com/amoeba/standardrb-action