Remove unused TimeRounder class
It appears this was used by a previous version of the implementation,
but usage was dropped on b66b99eeee2c3ca5cf8ef912461ba180fb37b96d:
"Submit measurements via the V2 API".

It is no longer referenced anywhere in the code, other than the requires
and its own spec that I also removed here.