Implement the logger tests using minitest
This is a sample of how tests could look like with Minitest, if we want
to go down that route of converting them.

It's a bit of a manual process, but this has the setup necessary to make
it work. Minitest also has a more spec-like syntax that's similar to
RSpec, if we wanted to use it, but this just goes with the most bare and
common Minitest usage.

I also ran into a bit of an issue with the ENV swapping usage/helper,
because of the callbacks nature of things it doesn't work the same way
here, so I switch that to swap the config instead. I guess that makes
sense, that the logger doesn't need to be concerned about the ENV var
but just the config, and we could test the ENV var elsewhere (e.g. in
the config tests.)