Implement logger tests using ENV config swapping like the original ones
I'm not sure if we should keep using ENV config for these tests here, or
if we should rely on our own controlled config and swap those values
instead. We can do the specific ENV testing on the config itself then.

For now, I want to simply keep the tests the same as before, so I'm
keep the extra describe block to be able to more easily wrap the env
setup like the previous `around` rspec version.