Convert measurement and middleware specs to minitest
There's some weird connection between updating the ENV and making sure
the config reflects that, which required me to make sure the config gets
reset & cached again when the ENV is swapped. I think we should stop
using ENV swapping in most of these tests, in favor of config swapping,
and only leave ENV swapping to test config setup, to simplify this stuff.