Ensure the store is always reset after each test that may touch it
Otherwise we can end up with data leaking in the singleton store
affecting other tests, depending on the order these tests are executed.

This was reproducible with the following seed:

    bundle exec rake TESTOPTS="--seed=46035"

consistently resulting in the following error:

      1) Failure:
    Judoscale::WorkerAdapters::Que::#collect!#test_0001_collects latency for
    each queue [./judoscale/judoscale-ruby/test/worker_adapters/que_test.rb:38]:
    Expected: 2
      Actual: 3

    42 runs, 101 assertions, 1 failures, 0 errors, 0 skips

Ensuring the store is always cleared after these specific tests prevents
the error. It might be nice to circle back on this and add a general
clear after each test run to make sure it's always in a clean state, not
leaking data that may affect other tests.