Quote Ruby versions to avoid YAML float conversion issue with 3.0
Due to an issue with how `3.0` is parsed as float 3 in YAML [1], the
workflow actually started building with 3.1 once it was released a
couple weeks ago during Christmas.

The current workaround seems to be quote it to avoid the float
conversion, so for consistency I quoted all of the versions.

I'm going to look into including 3.1 in the build matrix, but as a
separate commit since it seems to require code changes.

[1] https://github.com/actions/runner/issues/849