Fix new code lint warnings raised by latest standardrb
`rubocop-performance` added a new `Performance/StringIdentifierArgument`
check in version 1.13.0, which `standardrb` started warning about in
version 1.6.0.

https://github.com/rubocop/rubocop-performance/blob/v1.13.1/CHANGELOG.md#1131-2022-01-01
https://github.com/testdouble/standard/blob/v1.6.0/CHANGELOG.md#160

There's one other place that standardrb tried to correct but it fails
tests if we do:

    Object.const_defined?("Judoscale::Config")

trying to turn the above string into a symbol causes the following:

    wrong constant name Judoscale::Config

Instead I think we can remove the conditional check entirely, since the
Config should always be defined when requiring the lib.