Split ENV setup & restoration so we can rely on minitest hooks
Minitest doesn't provide an "around" hook which made me use a
combination of `def setup` with `super` to workaround some of that and
be able to override the ENV vars for some tests, but that combination
with the other before/after hooks provided by minitest spec's didn't
really seem to work well together. This is a different take on it: split
the setup & restoration so we can use before/after hooks.