Remove exception handling/reporting to the API in favor of logging only
Reporting exceptions through the API was meant to help investigate
issues, but it appears to almost never been useful, and possibly a
source of bugs, so we're removing this handling, and simplifying the
logic to just catch any exceptions and log them with backtraces to the
currently configured logger, so they show up in the client logs.

Having them in the logs will still be helpful to diagnose issues if they
happen, since it will allow clients to grab them as needed. Moreover,
the exception rescuing is necessary due to how Threads are handled in
Ruby: raising inside a thread that is not the main one would by default
terminate it and report its exception to STDERR, but we need the thread
to continue running and reporting scale metrics in this case. So we
catch the exceptions, log them for further analysis if needed, while
keeping the reporter thread up.