Allow logging multiple messages in separate lines, all tagged
Instead of logging multiple lines separately to tag them individually
and facilitate searching for all logs related to Judoscale, we're going
to group the logging of related messages together (e.g. exception
messages & backtraces), so they're sent to the underlying logger as a
single log call, which should ensure they're logged as a single "unit"
and are output together with the same timestamp in the logs.

These individual lines are going to be tagged so they are easy to search
for if needed.