Make measurement metric a required argument, move it upfront
The metric is an identifier of the value we're reporting, we're making
it a required argument of any value we push to the store for reporting
so that it's included with the reporting sent to the API, and it's
clearer from the caller what that measurement is about, instead of
inferring it's "queue time" based on the absence of the metric name.
(which was the case for both web and some of the worker adapters before,
but not all of them.)

To make that even more prevalent, reorder the arguments of the
Measurement object to make metric come first. This doesn't change the
reporting though, where metric is still last.