Move request logging to the middleware
Avoid logging as a side effect of calculating queue time, instead log
only when we're using the queue time and pushing it to the store (i.e.
collecting the queue time for reporting), which makes it more similar to
how the worker adapters implement logging.

This required a bit of refactoring on the request object to expose the
request ID & size, so we could keep including that information in the
debug request log.

It also includes a new test for this particular request logging
functionality.