Setup the Rails logger in the tests with a StringIO instance
This allows us to test the logs more easily since we will always have
instance to a `log_string`, that is automatically reset/cleared after
each test.

It also removes the need to be swapping the configured logger in
specific tests where we needed to test logging, like the reporter.

Note: I did not touch the logger tests themselves which also swap the
configured logger, because those seemed specific to that code.