Use a regexp accepting a possible ms variation to avoid random failures
We're using delta assertions for most of these queue time related tests
because we're not stopping time, so the log check must follow a similar
rule otherwise we run the risk of random failures due to time variation.

We got one on the CI that helped identify this issue: (truncated some
output to facilitate readingj)

    1) Failure:
    Expected "..., [2022-01-26T21:41:16.237102 #1600] DEBUG -- : [Judoscale] Request queue_time=5001ms network_time=0ms request_id=req-abc-123 size=5\n"
    to include "Request queue_time=5000ms network_time=0ms request_id=req-abc-123 size=5".

This indicated the time variation issue making this a test bound to fail
randomly, which we're hopefully fixing here by using a regexp that
accepts some level of milisecond variation there.