Check for worker adapter enabled upfront to make sure it is required
Both Sidekiq and Rescue try to require the respective libraries when we
check if they're enabled, if we don't do that in the test setup,
depending on the order in which tests run, these new logging tests might
randomly fail because the constants aren't defined yet.

This was reproducible with the following seed that happened on CI:

    % be rake TESTOPTS="--seed=61479"

    ...E.....................................................

    Finished in 0.302106s, 188.6755 runs/s, 473.3438 assertions/s.

      1) Error:
    Judoscale::WorkerAdapters::Sidekiq::#collect!#test_0004_logs debug information for each queue being collected:
    NameError: uninitialized constant Sidekiq

Performing an "enabled?" check in those tests fix this issue, and it's
how other tests are also being setup currently.

This should no longer going to be a problem once we move away from the
"enabled?" checks in the future.