Test config + ENV vars separately, use config overrides for other tests
The config is the only piece that really knows about ENV vars currently,
reading them and setting internal properties as needed. All other pieces
of the code interact directly with the config instead of reading from
the ENV.

Reflect that relationship and encapsulation within the tests as well,
by overriding config values for different test needs to drive behavior,
and leaving the config tests to handle how ENV vars affect the config
object.