Allow worker adapters to be configured directly and add tests
The worker adapters are converted during init to a list of instances of
the adapter classes (which are all singletons), so setting them via
configuration block wasn't going to work unless we ran that through the
same process.

This now allows the worker adapters config to be set during
initialization instead of ENV vars only, and adds tests that cover that
and all other configs being setup via a block.