Drop support for most ENV vars in favor of the new config style
Judoscale will no longer support ENV vars for most configurations,
instead the documentation will suggest people to create a config
initializer to tweak the various config options available.

One exception to that is going to be the JUDOSCALE_DEBUG, or `debug`
config option. We want customers to be able to set that on the fly
without having to setup a config file or deploy the app, since that's
been very helpful to support / troubleshooting purposes, so this ENV var
will continue to be supported.