Lazily load adapters when the reporter start
Instead of eagerly loading adapters when the config is set, move the
loading for the reporter when its starting. This prevents us from
eagerly attempting to load all default adapters upon config
initialization, before the client code had a chance to run its own
configure init block to override those defaults.

Now we'll just store the list of worker adapters with the config, and
try to load them later on when the reporter starts. This loading
mechanism might change again when we split those worker adapters, but
for now this allows the config to avoid loading them all upfront,
working more as intended.