Move the "queues" property to the base worker adapter
All of them shared the concept of a queues list that started with a
"default", but some of them had it implemented as a Set, others as an
Array. This merges that implementation moving it to the base worker
adapter class so they can all share, and uses Set as the representation
for that list, as there's really no way duplicate queue names could
exist for each adapter.

Minor tweaks to the adapter code were made, the main one was on Sidekiq
that started using the |= operator instead of overriding the whole
queues list. I could have set that wrapping with a `Set.new()`, but
moving to this implementation doesn't seem to cause any harm, other than
the fact that the "default" queue will always be there (like the other
providers) and come first in the list, instead of possibly last, because
of the order in which we collect those in the Sidekiq adapter
implementation.