Refactor the check for number of queues exceeding the limit
Share this implementation across worker adapters since they all
implemented it similarly.

This required us to create a way to extract the adapter name from the
class name so we could use a single logger call / statement that
contains the adapter name in it, but that might come in handy as we move
more code around in the future.