Fix incorrect args with track long running jobs on sidekiq adapter
It looks like when I implemented the measurement changes to require the
metric identifier, and moved that arg to be the first when pushing to
the store, I missed this spot on sidekiq, which wasn't covered by tests.

This fixes the arguments order, and includes some tests for that logic
to help prevent breaking it in future refactorings.