Create a queue filter configuration available for each adapter
This allows the configuration of a custom queue filter proc that will
run for each queue name, allowing even more control over which queue to
collect metrics for reporting.

The filter proc must return a "truthy" value to include that queue in
the collection, any "falsy" value will exclude it.