Setup a default queue filter for UUID formats
In general queues named with a UUID format are dynamically generated,
for example in multitenant scenarios where each user/account gets its
own queue, it's common to name those queues after an UUID identifying
the user/account. That can lead to hundreds/thousands of queues and we
don't want to be collecting that many, so the default filter will
exclude those.

This filter is configurable and will override the default UUID one, so
it's still possible to bypass it if the client has a real need for it.