Report up to `max_queues` instead of dropping all of them
We will no longer skip reporting if the queues size is higher than the
configured `max_queues`, instead we'll report up to that configured
number of `max_queues` so that these reports start showing up in
Judoscale, skipping the rest of the queues.

In order to determine which queues to keep and which to drop, instead of
relying on whatever sort order we got those, we decided to always sort
them by length of the queue name. This should keep the list of queues
collected somewhat consistent, while also more likely avoiding metrics
reporting for longer queue names that might be dynamically generated.

`max_queues` is still a configuration option so anyone will have the
ability to tweak that number if they need to.