Cleanup after tests run instead of as part of the setup
Try to leave no trace behind after each test is executed, as a cleanup
step, instead of as part of the test setup. The setup phase should be
more about setting up the world, in case of the DB we should be able to
"trust" it starts up empty, and if that's not the case we should ensure
that before the suite runs instead.