Combine `debug` and `quiet` configs into a single `log_level`
This allows us to simplify the logging implementation via a
configuration to control the log level. Our logging wraps the underlying
logger (i.e. Rails.logger) and only passes down messages that have a
matching severity, dropping lower severity messages on the floor.

If the `log_level` isn't set, we simply pass everything down and allow
the underlying logger to handle the proper log level.

The `debug` level is treated a bit differently though, as we want to be
able to let clients enable `debug` temporarily, even if their apps are
setup with a higher level like `info`. In that case, we log at the
logger level instead of our configured level, to ensure it gets logged,
and prefix all messages with an extra `[DEBUG]` tag, to make them easier
to identify and search for.