Separate the report concern from the metrics store
The report instance is constructed by the reporter now when it's about
to send the metrics, by flushing the store to gather all collected
metrics and passing that down to the report, allowing the store to be
more decoupled from the report, i.e. making them both talk in terms of
"metrics" but not knowing each other directly.