Move the report metrics API to use JSON instead of CSV
This simplifies the reporting API and makes it more consistent with the
registration by always using JSON to talk to the API.

The API used a combination of query params for some attributes, and a
CSV body. This merges the whole payload into a single JSON body. We are
now only sending the `dyno` and dropping the `pid`, as that has not been
very useful information.

We still build the metrics as a list of arrays instead of JSON objects
to keep the reporting simpler and the payload smaller, which makes it
still very similar to a CSV from the looks of it, but it's just one
piece of the JSON payload, and enables us to extend it further in the
future.