Start refactoring the reporter into the "collectors" interface
This removes the "store" concept from the reporter, in favor of
initializing it with a list of "collectors" to collect metrics from. On
each reporter loop it will simply go through each collector and ask to
collect metrics, and report those metrics.

For jobs, it will collect metrics at that moment. For the web, it will
simply flush the already stored metrics from web requests. But the
interface for both is going to be essentially the same, simplifying the
reporter and allowing reporters for different dyno types to be created
later. (for example a worker dyno reporter doesn't need to be concerned
with the web metrics.)