Create a "job collector" interface wrapping our current worker adapters
Eventually we want to make each worker adapter essentially a
"collector", but for now since we have that implementation already in
place, this allows us to wrap them with a new job collector to keep the
interface the same across all collectors, allowing us to further
refactor and simplify the reporter.

It also enables us to decouble the reporter tests from the worker
adapters, since all we need is a "collector" interface.