Add basic tests for web & job metrics collectors, fix job collection
We can't just use an Array for the job collection yet because `push`
adds each element individually, and we want them to be considered a
single "metric", so without changing the worker adapters implementation
(which I'm trying to avoid in this first pass), we need to wrap that
into our own object to track metrics.

I can't use the MetricsStore because it's shared with web, it's only
meant to be used there as a singleton, so creating a simple wrapper for
it for now.