Disable Ruby 2.6 on the Rails lib for now
Ruby 2.6 doesn't work with Rails 7, so the build fails with that
combination.

The Ruby lib was using Rails 6.1 due to other constraints, so it works
there, but not here if we allow installing/running against Rails 7.

Later I want to run the tests against different versions of Rails, but
this will do for now to get a green build.