Set the Rails logger via a Rails initializer
Move away from checking for Rails being defined in the Ruby library in
favor of initializing it via the Railtie.

This doesn't work with our current `Config#reset` setup in tests,
because it doesn't allow to be reset, so I removed the support loading
from the Ruby library that was causing a conflict there, but for now
this should be okay as we have very little logging in the Railtie
anywya. If we need to test logs later we can reorganize this config to
be a module within Rails and workaround this small limitation.