Set the logger from the `Rails.logger`
If the application does not have a configured logger, the `Rails.logger`
will be initialized by default, and `config.logger` will be `nil`, so we
need to actually reference `Rails.logger` directly.

Our test app does set the `config.logger` so it wasn't catching it. I'll
think of a way to test the `nil` case, but this should at least allow
our sample app to work again now.