Add initial infrastructure for registering an adapter
We adapter needs to collect some information for registration, so they
need to implement an ".adapter_registration" method that we can use to
collect that information during the registration process.

I am thinking about pushing that up to the register class, maybe that'd
be okay? I was a bit afraid of having to store that to retrieve later
during registration, since it'd happen at load time instead of runtime.