Allow each adapter to configure its metrics collector
Instead of by default running with the Web metrics collector, let each
adapter set one as part of the registration, which allows the Rails
adapter be responsible to configuring the Web metrics collector when
it's loading, instead of assuming there is one.

Also let the metrics collectors themselves be responsible for
determining if they should run their collection based on the context
they are in, so that jobs collectors can decide only to run on the first
dyno to avoid redundant collections on other dynos (as they'd likely be
the same metrics anyway.)

This should be the same behavior as before, except we're not loading the
existing worker adapters anymore from here on out, as we need them to
become adapters using the new implementation going forward. That's going
to be a separate piece of work, but this sets the stage for that to
happen next.