Reorder arguments to build Metric objects
We push metrics arguments to the store with value / time, but build the
actual Metric objects that we store with these arguments reversed, i.e.
time / value. This changes the order of the Metric arguments to keep the
Store & Metric more consistent. It is mostly an internal change that
shouldn't affect callers / users of the code in general.