Initial implementation of the Sidekiq job metrics collector
The collect implementation comes straight from the previous worker
adapter (to be removed), and the helper methods around it come from the
base class.

The plan is to move them back to the base JobMetricsCollector
eventually, but I need to figure out how to access the configuration in
a shared way, and how to create it on-the-fly since right now it's using
the previous "default" list, not the registration as a hook to create
it.