Rename register => add in the context of adapters
We already have a use for the word "register", which is registering the
adapter with the Judoscale API, so this renaming is to prevent us
overloading that word with a different kind of "registration" within the
code base.

This is all subject to change as we solidify the implementation more,
even possibly chaning how the "reporter registration" happens.