Send an initial report to notify the Judoscale API
Since we won't have a registration step anymore, send an initial report
request to the Judoscale API as soon as the reporter thread starts, but
before waiting or collecting any metrics.

This effectively "replaces" the need for registration as we'll still
attempt to hit the API as soon as the reporter starts, with all the
context/metadata about the reporter and its environment.