Remove registration, include metadata with reports
Simplify the reporter logic by always including contextual metadata with
each report, and removing the initial registration step which included
such metadata.

The metadata is important for support/troubleshooting/debugging
purposes, and keeping it on every report to the API keeps these reports
more self-contained in general, and simplifies the code considerably.