Fix ruby tests & sidekiq lint to get the build green
The Ruby library tests are failing due to the job collector base
implementation logging by default when initialized, requiring an adapter
name, which didn't exist for the test job collector we were creating for
the reporter tests... for now we're using a placeholder of the sidekiq
version (which still exists there), but that should be removed when this
is final and we have the config stuff properly working.

The lint was failing with a useless variable assignment.