Move metrics reporting upfront when starting the reporter
This allows us to send an initial report to the API to let it know the
reporter is working as soon as it is started, while also including the
first batch of job metrics collected at that moment. There should be no
web metrics to include for the most part, but at least job metrics can
be reported right away.

We don't foresee issues with running the collection as soon as the
reporter start, even if we start a reporter on boot later, it should be
fine to start collecting metrics during initialization then.