Let adapters add their "config" to the main judoscale config module
This works similarly to the adapters adding themselves to the list of
known adapters, but for configuration that each adapter may want to
expose.

Right now only job adapters have configs and they use the same base
config class, but we want them to be responsible for telling the main
config about it, rather than the other way around, as all the pieces
about their adapter implementation should live in their own code.