Start a reporter after the Rails app is done initializing
This way we can already kick an initial metrics collection & report to
let Judoscale know the reporter is working, without having to wait for
any requests to hit the app.

It also means worker-only apps running with the Rails adapter get their
own reporter, without needing to run a web dyno or anything.

Worker-only apps without the judoscale-rails adapter still don't have a
way to run reporters, that's something we'll look into later, but this
should already cover most usage.

The way it's implemented is by running a new worker every time the
process PID changes, so that if you're running a single thread server
(e.g. Puma in thread mode only), you only get this initial reporter
running, and it will be responsible for all collection & reporting.
In cluster/workers mode, however, a reporter will be created during
boot, and a new one for every forked process whenever a new request hits
the request middleware, so each process get their own reporter thread.
This is necessary because each process keeps an in-memory "store" of
request metrics to flush every few seconds, and that store isn't shared
among different processes, so we have each process run its own reporter
thread and report the stored metrics on their own, a simple approach to
make sure all processes are always covered, and that if a process is
killed and restarted (e.g. by puma or unicorn), a new reporter can start
again automatically to handle that new process as well.