Make config a default argument on reporter & request metrics
Since it is a singleton global, we can simplify the callers a bit by
removing the need to pass it around manually, relying on the default
argument pointing back to the singleton instance instead.

We still get the flexibility of passing the config if we need,
especially for testing purposes, but from the callers perspective it is
much simpler.