Make all log levels behave the same consistencly
If we don't have a configured log level, we'll just let the underlying
log level dictate whether the logging should really happen, no changes
here.

However, when there is a configured log level, we first determine if we
can proceed with logging at that level, otherwise we'll drop the message
on the floor. In other words, we don't log DEBUG if the configured log
level is INFO for example.

If we're allowed to log at the configured level, we'll just determine
if the underlying logger is allowed to log at that level too, to let it
through; otherwise it means we have enabled a lower log level than the
underlying logger, but we still want to ensure those logs are going to
the output, and we do that by prefixing/tagging them with the level,
e.g. `[Judoscale] [DEBUG]`.