Create global web & job test adapters for the judoscale-ruby test suite
Move the test adapters previously only used within the reporter test to
be available for the whole test suite. Having them setup globally allows
us to rely on a set world that contains a common scenario for testing
both the reporter and the configs, which are provided by the base
judoscale-ruby library.

It will also make it easier to extract the other adapters as we don't
need to rely on them so much for various config tests we were doing
before, since we now have a test job config readily available there.