Lock Active Record on v6.1 due to Ruby 2.6 compatibility
We need to setup a matrix that can run different versions depending on
the current requirements, e.g. for Ruby 2.6 we need AR 6.1 since 7.x
doesn't support it anymore, and to create that sort of configuration we
need different Gemfiles. For now I'm trying to stay away of that setup,
and this guarantees a green build across all Ruby versions.