Respect explicitly disabled job adapter from collecting metrics
We were allowing job adapters to be disabled via configuration, but were
not honoring that anymore, I believe when moving from the worker
adapters that filtered those to the new job adapters infrastructure.
This ensures we respect the configuration and adds better tests to cover
it via the reporter.

Note: we're still logging that all judoscale adapters are running, but
that is not really true when you disable one job adapter via this
config. I'm looking into that separately.