Add a root page for the rails sample app, and link to request catcher
Add a small bit of copy to both Rails & Sidekiq sample apps pointing to
the request catcher URL, to make it easier to get there once we boot up
the app.

It might be too late if you're looking to test the initial reporting,
but for the most part it should make it easier to get there as I keep
having to lookup the URL as I'm booting up the apps to test.

Plus having some inline instructions feel nice there for anyone looking.