Move Que tests to postgres
Create and drop the test database for each run, and migrate Que using
its own migration function to ensure we have the most up-to-date tables
setup for our testing.

Que is really meant to be used with postgres, so using it for our tests
will allow us to have more realistic tests and possibly fix any postgres
specific issues with Que usage.