Remove the collection of a "default" queue across all job adapters
This was a partial fix to ensure the adapters would report something for
a "known" queue every time they started, because some adapters aren't
able to keep track of previously known queues (mostly the SQL based ones
that don't keep jobs around).

The backend however is able to handle those gracefully by assuming a "0"
value if it is not getting reports of a previously reported queue, so it
doesn't really need the "default" hard-coded implementation we had here.
Plus, it'd only work for the "default", not for other custom queues.