Test judoscale-rails against both Rails 7 and 6.1
Rails 7 doesn't work with Ruby 2.6, so we initially configured the test
matrix excluding that Ruby version to be able to test against Rails 7.
This new setup allows us to include 2.6 back, and run it (and other Ruby
versions) against Rails 6.1 as well, giving us more coverage that things
are working as expected across multiple versions of Rails going forward.

I chose not to include the .lock file for this additional Gemfile, as
those dependencies tend to be more stable and they are not the ones we
usually test against locally, but more for the build, so if something
happens to fail there with updated dependencies in this case, I think
it's usually better to know. (but again, that seems less likely to
happen with previous versions.)