Move Que sample app to SQL schema to setup DB correctly
The Ruby schema does not dump many of the necessary functions required
for Que to work correctly, so they require the app to use the SQL
structure file instead:
https://github.com/que-rb/que#additional-rails-specific-setup

For some reason I didn't run into this when initially creating the
sample app, and with my existing DB it was working, but trying to
recreate the DB from scratch failed. This fixes it.