Test judoscale-que with version 2 beta
In order to make sure we're prepared and green for an imminent version 2
release, add a separate Gemfile for the 2 beta version to the Que
matrix to start running tests against it now. Our sample app was already
using the 2 beta version and appears to be working fine, we just weren't
testing against it fully.

The `job_schema_version` field is now "not null", so we need to also
provide it when inserting into the table for testing. It already exists
in version 1, just wasn't required before.
https://github.com/que-rb/que/commit/6cc40bd089977f06eef27274b52e28907abbd745

Que v2 beta also requires Ruby 2.7+, so we excluded it from running
against Ruby 2.6 for now.
https://github.com/que-rb/que/commit/3f618ebe5dd200853dba2457af97a383c8796ea5